fix: resolve clippy warnings for strip_prefix, cast, and const assertions

+++ src/services/lockout_service.rs
@@ -121,7 +121,7 @@ impl LockoutService {
                 .set_ex(
                     &lockout_key,
                     locked_until.to_string(),
-                    self.lockout_duration as u64,
+                    self.lockout_duration,
                 )
                 .await?;
